MAKE Accessors trait with properties & methods that are not dependent to other PdfExporter methods
