+++ src/Pdf/Utils/PdfExporter.php
@@ -12,6 +12,7 @@ use Sfneal\Helpers\Strings\StringHelpers;
 class PdfExporter
 {
     // todo: make dispatchable
+    use Accessors;
 
     /**
      * @var Options
@@ -22,21 +23,6 @@ class PdfExporter
      * @var Dompdf
      */
     private $pdf;
-
-    /**
-     * @var string|null AWS S3 file path
-     */
-    private $path;
-
-    /**
-     * @var string|null AWS S3 file URL
-     */
-    private $url;
-
-    /**
-     * @var string|null
-     */
-    private $output;
 
     /**
      * @var View|string
@@ -127,35 +113,5 @@ class PdfExporter
     public function download(string $filename = 'output.pdf')
     {
         $this->pdf->stream($filename, ['Attachment' => true]);
-    }
-
-    /**
-     * Retrieve the PDF's output.
-     *
-     * @return string
-     */
-    public function getOutput(): string
-    {
-        return $this->output;
-    }
-
-    /**
-     * Retrieve the PDF's AWS S3 path.
-     *
-     * @return string|null
-     */
-    public function getPath(): ?string
-    {
-        return $this->path;
-    }
-
-    /**
-     * Retrieve the PDF's AWS S3 url.
-     *
-     * @return string|null
-     */
-    public function getUrl(): ?string
-    {
-        return $this->url;
     }
 }
